services/compliance_agent/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-pubsub==2.19.0
langchain-google-vertexai==0.0.5
langgraph==0.0.26
langchain-core==0.1.23
pydantic==2.6.1
